requirements.txt
# Python dependencies for docx2app - Document to JSON Converter
# Install with: pip install -r requirements.txt

# Core document processing
python-docx>=0.8.11

# Image post-processing (auto-crop whitespace, resolution limiting)
Pillow>=10.0.0

# Optional: For future enhancements
# lxml>=4.9.0
